requirements.txt
asgiref==3.3.0
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.4
cryptography==3.0
defusedxml==0.7.0rc1
diff-match-patch==20200713
Django==3.1.3
django-environ==0.4.4
django-import-export==2.4.0
django-phone-field==1.8.1
django-progressive-web-app==0.1.1
django-sendgrid-v5==0.8.1
django-shopping-cart==0.1
django-simple-referrals==0.2
django-treebeard==4.3.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
future==0.18.2
idna==2.10
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==1.3.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-http-client==3.2.7
python3-openid==3.2.0
pytz==2018.7
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==6.4.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
starkbank-ecdsa==1.0.0
tablib==2.0.0
timedelta==0.0.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
